Redesign the messaging system to support Firefox quirks
Firefox doesn't seem to properly support re-entrant messages, that is messages that sends messages, when they also make use of the response API. Instead of correctly waiting for the response and fulfilling the promise, the promises of the entire chain of messages get fulfilled instantly with undefined values.

This solution emulates the response API by opening a separate reply channel for messages.