Manually handle the default device
Firefox doesn't include the default audio output device in the device list, meaning that once a device is picked, it is impossible to go back to the default behavior and disable AuRo. This commit removes the default device from the list on supported browsers (like Chrome) and reinjects it manually to ensure it is always present.