Redesign the injectable code to support firefox
Firefox and Chrome have completely different broken behaviors when it comes to setSinkId. Firefox has a bug that prevents setting the sink when the media isn't playing, while Chrome refuses to set the sink while the media is playing. This means we need to have two different implementations that tries to set the media in the correct state before calling setSinkId on them.

This new version also have a better handling of updating the sink while paused though independent DOM element detection and orphan element detection.